Debugged Hosting a bunch
Also reorganized code more similar to Wallet

Also made some formatting changes to Wallet's code
