+++ app/plugins/Hosting/index.html
@@ -55,6 +55,7 @@
 		</div>
 		<div class='host-message' id='hmessage'>Estimated Competitive Price: 4500 SC / GB / Month</div>
 		<!-- JS -->
+		<script type='text/javascript' src='js/helpers.js'></script>
 		<script type='text/javascript' src='js/hosting.js'></script>
 	</body>
 </html>
