+++ app/plugins/Wallet/js/helpers.js
@@ -1,5 +1,6 @@
 'use strict';
 
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~ Global Variables ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Library for communicating with Sia-UI
 const IPC = require('ipc');
 // Library for arbitrary precision in numbers
@@ -14,6 +15,7 @@ var currentHeight;
 // Keeps track of if the view is shown
 var updating;
 
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~ Helper Functions ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // DOM shortcuts
 function eID() {
 	return document.getElementById.apply(document, [].slice.call(arguments));
